src/Services/BizFlow.AI/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
pydantic==2.5.2
chromadb>=0.5.0
requests==2.31.0
python-dotenv==1.0.0
google-generativeai>=0.8.3